backend/requirements.lock.txt
django==4.2.13
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.3.1
pymongo==4.7.3
python-dotenv==1.0.1
scikit-learn==1.5.0
numpy==1.26.4
gunicorn==21.2.0
whitenoise==6.7.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
rapidfuzz==3.9.6
langdetect==1.0.9
PyJWT==2.9.0
email-validator==2.2.0
